+++ backend/elements/Participant.py
@@ -1,29 +1,16 @@
-from elements._elementBase import ElementBase, docDB
+from noapiframe import docDB
+from noapiframe.elements import UserBase
 
 
-class Participant(ElementBase):
-    _attrdef = dict(
-        admin=ElementBase.addAttr(type=bool, default=False, notnone=True),
-        name=ElementBase.addAttr(default='', notnone=True),
-        login=ElementBase.addAttr(type=str, unique=True, default=None),
-        pw=ElementBase.addAttr(type=str, default=None),
-        seat_id=ElementBase.addAttr(type=str, unique=True, fk='Seat')
-    )
+class Participant(UserBase):
+    UserBase._attrdef['name'] = UserBase.addAttr(default='', notnone=True)
+    UserBase._attrdef['seat_id'] = UserBase.addAttr(type=str, unique=True, fk='Seat')
 
     def validate(self):
         errors = dict()
         if self['seat_id'] and not docDB.exists('Seat', self['seat_id']):
             errors['seat_id'] = {'code': 70, 'desc': f"There is no Seat with id '{self['seat_id']}'"}
         return errors
-
-    @classmethod
-    def get_by_login(cls, login):
-        result = cls()
-        fromdb = docDB.search_one(cls.__name__, {'login': login})
-        if fromdb is not None:
-            result._attr = fromdb
-            return result
-        return None
 
     @classmethod
     def get_by_seat(cls, seat_id):
@@ -40,7 +27,7 @@ class Participant(ElementBase):
     def delete_post(self):
         from elements import Session
         if self['_id'] is not None:
-            for s in [Session(s) for s in docDB.search_many('Session', {'participant_id': self['_id']})]:
+            for s in [Session(s) for s in docDB.search_many('Session', {'user_id': self['_id']})]:
                 s.delete()
 
     def offboard(self):