+++ backend/endpoints/__init__.py
@@ -1,7 +1,4 @@
-from .element import ElementEndpointBase
-from .login import LoginEndpoint
 from .system import SystemEndpoint
 from .switch import SwitchEndpoint
 from .onboarding import OnboardingEndpoint
-from .setting import SettingEndpoint
 from .participant import ParticipantEndpoint