+++ backend/helpers/docdb.py
@@ -1,3 +1,0 @@
-from helpers.mongodb import mongoDB
-
-docDB = mongoDB()